+++ ReleaseNotes.md
@@ -1,8 +1,16 @@
-# Release Notes for Fluid Infusion 4.0.0
+# Release Notes for Fluid Infusion 4.1.0
 
 [Fluid Project](https://fluidproject.org)
 
 [Infusion Documentation](https://github.com/fluid-project/infusion-docs)
+
+## What's New in 4.1.0
+
+* Maintenance release which resolves
+  * [FLUID-6730](https://issues.fluidproject.org/browse/FLUID-6730) Removed some obsoleted utilities - fluid.accumulate,
+    fluid.contains, fluid.add, fluid.stableSort
+  * [FLUID-6727](https://issues.fluidproject.org/browse/FLUID-6727) Improve documentation (retrospectively) about changes
+    in DataSource API in the 3.0.0 release
 
 ## What's New in 4.0.0?